Um, experimenting a lot
So apparently we have to make a version of the robot that path planner can control. We also have to do a bit of translating. Actions that aren't the drive base should be similar to what we already have thankfully. I am planning to look around online to see if we can find viable code that makes sense.